+++ R/calcInteger.R
@@ -31,6 +31,8 @@
 #'
 #' @examples
 #' copykit_obj <- mock_bincounts(ncells_diploid = 0)
+#' copykit_obj <- runVst(copykit_obj)
+#' copykit_obj <- runSegmentation(copykit_obj)
 #' copykit_obj <- calcInteger(copykit_obj, method = "scquantum")
 
 calcInteger <- function(scCNA,